Made a BigDecimal version. It starts having effect after about 100 games

+++ glicko/src/main/java/solutions/desperate/glicko/domain/model/Player.java
@@ -1,27 +1,29 @@
 package solutions.desperate.glicko.domain.model;
 
+import java.math.BigDecimal;
+
 public class Player {
     private final String name;
-    private final double rating;
-    private final double rd;
-    private final double volatility;
+    private final BigDecimal rating;
+    private final BigDecimal rd;
+    private final BigDecimal volatility;
 
-    public Player(String name, double rating, double rd, double volatility) {
+    public Player(String name, BigDecimal rating, BigDecimal rd, BigDecimal volatility) {
         this.name = name;
         this.rating = rating;
         this.rd = rd;
         this.volatility = volatility;
     }
 
-    public double rating() {
+    public BigDecimal rating() {
         return rating;
     }
 
-    public double rd() {
+    public BigDecimal rd() {
         return rd;
     }
 
-    public double volatility() {
+    public BigDecimal volatility() {
         return volatility;
     }
